+++ src/main/java/be/baur/sds/common/NaturalInterval.java
@@ -11,7 +11,7 @@ package be.baur.sds.common;
  * a..*	Right unbounded, but a..Integer.MAX_VALUE in practice
  *   a	Degenerate, equivalent to a..a
  *   
- * where a >= 0 and b >= a
+ * where a &gt;= 0 and b &gt;= a
  * </code>
  * </pre>
  *