+++ RELEASE_NOTES.md
@@ -2,7 +2,7 @@
 
 This release introducdes a few refactoring changes with limited impact.
 
-## [2.1.0] - 2022-??-??
+## [2.1.0] - 2022-12-01
 - `Added` abstract superclass MixedType for types with simple content.
 - `Changed` Content Enum moved to package be.baur.sds.common.
 - `Removed` (Natural)Interval constructors, added factory methods.